Only handle rspec result from the same test file
Rspec have this feature call shared examples which causes the plugin to
break because it was running other tests in a different file. It causes
the line value outside of range bug as the current test file might not
contain the same number of line as the shared example test.